requirements.txt
requests
shodan
python-dotenv
rich
